Update: Fixed GoodPut Graph for UDP
Fixed the graph for the throughput in the application layer, it looks almost the same as ThrRx but with lower values because it doesn't have the phy layer overhead. The values are resampled and currently only graphs for one Node, so in case of multinode (or multiUE) configuration it has to be updated.