Update: Changed inner workings of Cqi
Changed the way the cqi algorithm is chosen, added a function calles SetCqiModel that defines which algorithm to use, the advantage is that in the future if we want to add a new algorithm it can be done easily. Created an enum with all the algorithms including the default one, it is important to note that SetCqiModel overrides Set, so if we set the default we can still call the set function. I tested the default algorithm and the ProbeCqi algorithm and both worked. The ProbeBLER algorithm HAS NOT been implemented yet.