Add: script to run sims in parallel
To run it one must configure in parameters the different simulations to run, there is some info at the begging of the script that is important and currently it doesnt implement a nice way to check the progess of each simulation, but it send the stdout to a file. If montecarlo number is greater than 1, a folder where all the simulations will be stored is created. Usage is as simple as bash parallel.sh IMPORTANT: it calls the bash command for now, if you dont have it will try do add sh too