Add & Update: First try to separate distribution
First attempt to separate the setup of the physical distribution of the UE and gNbs, the idea is to have a separate file where the distributions are located and in the main .cc a switch case is used with choosable by an argument. Currently only the default distribution is implemented. For the future: in case multiple variables are needed, we can create a struct with them and pass that to the functions (I borrowed the idea from lena-lte-comparison-user.cc