Update: Added generic physical distribution graphs
Now physical distributions can be graphed almost automatically, all that is needed is call the correct function inside physical-scenarios.cc, the graphs arent completely ready yet so they need to be checkout in terms of sizing and dimensions. I moved all the existing scenarios to the left and down, so they are centered in x=30 (i thought it would make them more visible if we put them directly in the paper), the graph name was changed to "phyDistro.png" and the images were moved to the "images" folder.